alx_travel_app/requirement.txt
amqp==5.3.1
asgiref==3.9.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.2
decouple==0.0.7
defusedxml==0.7.1
Django==5.2.4
django-autocomplete-light==3.12.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dnspython==2.7.0
docopt==0.6.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mysql-connector-python==9.0.0
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pipreqs==0.4.13
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
rabbitmq==0.2.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
yarg==0.1.10
